requirements.txt
Flask==2.0.2
tensorflow==2.17.0
Pillow==9.5.0
roboflow==0.2.15
opencv-python-headless==4.10.0.82
